make storage.set work on intel HD5500, so tests pass ok

+++ src/lib/THClStorageSet.cl
@@ -1,6 +1,10 @@
-kernel void THClStorageSet(global float *data, long index, float value) {
+kernel void THClStorageSet(global float *data, int index, float value) {
   if(get_global_id(0) == 0) {
+//    int index2 = index;
+//    data[index2] = 44;
     data[index] = value;
+//    data[2] = index2;
+//    data[3] = value;
   }
 }
